Log in to last.fm and get credentials.
First step towards allowing Scrobbling. The whole last.fm PHP API has
been included here; this is somewhat unnecessary, as the only call
it will be making is the auth call when the user first logs in.
For the moment though, it suffices and I'm loathe to go through
at the moment and strip out everything that's not needed.

NB in this commit the radio won't work.