Don't use iframes for POST in JS last.fm API
Last.FM now have CORS headers in their API, which means that we can query the
API directly without using a proxy script. We still face a problem of signing
the request without exposing our API Secret.  I have solved this via
signCall.php which is called from lastfm.api.js whenever an authenticated call
is required, which fires a synchronous request to the host server to sign the
call and return the appropriate MD5.

The relevant section in the conditions of use of the Last.FM API is 4.4, where
we agree we will not make more than 5 requests from an IP per second -- not
that we won't make more than 5 requests from an API key per second. Other than
the orignal login call, all API calls are made from the client. Our server
serves an AJAX response for every such call made, but only to the client and
not the Last.FM server. We should easily fall under the required limit.