Scrobbles on 50% of track played.
A little unsatisfying that 'now playing' gets displaced by a
scrobble, but I guess that's the usual behaviour? I am not
experienced enough with the Last.FM API to say one way or the
other.

NOTE: This still technically violates the Last.FM ToS, in that
seeking is allowed. This should be dealt with...

Also doesn't seem to work correctly on Chrome any longer.
Neither does the original version by tburny, however. Either
Chrome or responses from Last.FM have changed, I suppose. I have very
occasionally had a track play in Chrome, so I currently suspect a race
condition of some sort, but that's a stab in the dark.